Update text colors in Other Projects directory/remove extra css.
The other projects directory text colors weren't update with the main project site, for darkmode and lightmode text colors, like I updated a few releases ago.

There was some extra in :root, :darkmode, that I removed which isn't required for the other projects directory, but the whole main.css in the other projects directory, still needs to be updated and cleaned.